requirements.txt
# Agno AI 投资分析系统依赖包

# 核心框架
agno>=0.1.0

# AI模型和API
openai>=1.0.0

# Web界面和API
fastapi[standard]>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0

# 数据处理
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0

# 金融数据
yfinance>=0.2.0
duckduckgo-search>=3.9.0

# 工具库
python-dotenv>=1.0.0
requests>=2.31.0

# 可选依赖
matplotlib>=3.7.0
seaborn>=0.12.0

# 开发工具 (可选)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 